feat: 기능별 중간 레벨 컴포넌트 분리 - AdminPanel 컴포넌트
