+++ src/basic/components/features/AdminPanel/index.ts
@@ -1,0 +1,1 @@
+export { AdminPanel } from "./AdminPanel";
